### 2024-07-15 #6991 @kyleniemeyer
Hi @jabirali, thanks for your submission. Due to the relatively small nature of your package, the JOSS editorial board is going to take a look to make sure it meets our requirements for [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) before proceeding to review. We should get back to you within a week or two at most.
