### 2024-08-09 #6991 @jabirali
Dear @phibeck,

Thanks a lot for agreeing to edit my submission! I have now had a look through the list of reviewers, and I believe the following referees (listed by their GitHub handle) might be relevant to my submission:

- alejandrojuria
- yw-fang
- kaelynj
- santoshkumarradha
- nmoran
- sabinomaggi

I didn't find any referees in the system with a background in the particular framework I've implemented, but the suggestions above all appear to have some background in condensed matter physics and experience with either tight-binding models or superconductivity. I therefore believe they should be qualified to judge my submission.

Kind Regards, Ali.
