### 2024-08-12 #6991 @sabinomaggi
Hi @phibeck, I guess I could do that, but I am on vacation now until the end of the month and will be attending a couple of conferences in the first half of September. Until then I could read the paper, but I will not be able to check things like installation or functionality of the package.
