### 2024-08-22 #7134 @editorialbot
[REVIEW]: Bodge: Python package for efficient tight-binding modeling of superconducting nanostructures
**Submitting author:** @jabirali (<a href="http://orcid.org/0000-0002-3725-0885">Jabir Ali Ouassou</a>)
**Repository:** https://github.com/jabirali/bodge/
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.1.0
