### 2024-09-01 jabirali/bodge#1 @mdavezac
Community guidelines
For publishing with JOSS (openjournals/joss-reviews#7134), bodge will require community guidelines. The exact format is not mandated AFAIK, but many authors include a CONTRIBUTING file, as per https://docs.github.com/en/site-policy/github-terms/github-community-guidelines.
