### 2024-09-01 jabirali/bodge#2 @mdavezac
JOSS nitpicks
In the context of openjournals/joss-reviews#7134:

- It would be good to mention the targeted platforms, e.g. Windows/Linux/MacOS machine, and that GPU and MPI platforms are not targeted.
- For the sake of less experienced pythonistas, the tutorial might include a statement about having to do `pip install matplotlib`
- perhaps mention that `plt.show()` is sometimes necessary to actually show the plots, depending on the context (e.g. okay in Jupyter, but less so when copy-pasting each tut to a file).
