### 2024-09-05 jabirali/bodge#1 @jabirali
Thanks for the suggestion! I have now pushed a [CONTRIBUTING.md](https://github.com/jabirali/bodge/blob/main/CONTRIBUTING.md) file to the `main` branch, please have a look if it looks clear or could be further improved.
