### 2024-09-06 5d2bf64
docs: mention architectures targeted in readme
(Thanks @mdavezac)
