+++ README.md
@@ -1,21 +1,44 @@
 # Bodge
-Bodge is a Python package for constructing large real-space tight-binding models. Although quite general tight-binding models can be constructed, we focus on the Bogoliubov-DeGennes ("BoDGe") Hamiltonian, which is used to model superconductivity in clean materials. In other words: If you want a lattice model for superconducting nanostructures, and want something that is computationally efficient yet easy to use, you've come to the right place.
+
+Bodge is a Python package for constructing large real-space tight-binding
+models. Although quite general tight-binding models can be constructed, we focus
+on the Bogoliubov-DeGennes ("BoDGe") Hamiltonian, which is used to model
+superconductivity in clean materials. In other words: If you want a lattice
+model for superconducting nanostructures, and want something that is
+computationally efficient yet easy to use, you've come to the right place.
 
 Research papers where this code has been used include:
 - [DC Josephson effect in altermagnets][2]
 - [RKKY interaction in triplet superconductors][3]
 - [Dzyaloshinskii-Moriya spin-spin interaction from mixed-parity superconductivity][6]
 
-Internally, the package uses a sparse matrix (`scipy.sparse`) to represent the Hamiltonian, which allows you to efficiently construct tight-binding models with millions of lattice sites if needed. However, you can also easily convert the result to a dense matrix (`numpy.array`) if that's more convenient. The package follows modern software development practices: near-complete test coverage (`pytest`), fast runtime type checking (`beartype`), and mostly PEP-8 compliant (`black`).
+Bodge can be used on anything from a normal laptop to an HPC cluster, as long as
+it runs an operating system where the SciPy stack is available (e.g. Linux,
+MacOS, or Windows). It is mainly meant for CPU-based calculations on one node
+(i.e. no MPI support), however some functions on the `development` branch
+contains optional support for GPU-based calculations via CuPy (CUDA).
+
+Internally, Bodge uses a sparse matrix (`scipy.sparse`) to represent the Hamiltonian,
+which allows you to efficiently construct tight-binding models with millions of
+lattice sites if needed. However, you can also easily convert the result to a
+dense matrix (`numpy.array`) if that's more convenient. The package follows
+modern software development practices: near-complete test coverage (`pytest`),
+fast runtime type checking (`beartype`), and mostly PEP-8 compliant (`black`).
 
 The full documentation of the Bodge package is available [here][0].
 
 ## Quickstart
+
 This package is [published on PyPi][4], and is easily installed via `pip`:
 
     pip install bodge
 
-Bodge should be quite easy to use if all you want is a real-space lattice Hamiltonian with superconductivity. For instance, consider a $100a\times100a$ s-wave superconductor with a chemical potential $μ = -3t$, superconducting gap $Δ_s = 0.1t$, magnetic spin splitting along the $z$ axis $m = 0.05t$, and nearest-neighbor hopping $t = 1$. Using Bodge, you can just write:
+Bodge should be quite easy to use if all you want is a real-space lattice
+Hamiltonian with superconductivity. For instance, consider a $100a\times100a$
+s-wave superconductor with a chemical potential $μ = -3t$, superconducting gap
+$Δ_s = 0.1t$, magnetic spin splitting along the $z$ axis $m = 0.05t$, and
+nearest-neighbor hopping $t = 1$. Using Bodge, you can just write:
+
 ```python
 from bodge import *
 
@@ -35,20 +58,40 @@ with system as (H, Δ):
         H[i, j] = -t * σ0
 ```
 
[…]
