### 2024-09-06 9a3789b
docs: add matplotlib note
(Thanks @mdavezac)
