### 2024-09-06 jabirali/bodge#2 @jabirali
Thanks for your input!

>  * It would be good to mention the targeted platforms, e.g. Windows/Linux/MacOS machine, and that GPU and MPI platforms are not targeted.

Good point, this is now added [near the top of the README](https://github.com/jabirali/bodge/blob/main/README.md?plain=1#L15-L19) on the main branch. (I also added to the README that there is some _optional_ CUDA support on the development branch.)

>  * For the sake of less experienced pythonistas, the tutorial might include a statement about having to do `pip install matplotlib`

Good catch! I used to have `matplotlib` as a dependency of the project, but recently removed this to slim down the package. The docs still use `matplotlib` so it's good to mention explicitly.

I have added a note about this under ["Getting started" in the tutorial](https://jabirali.github.io/bodge/tutorial.html#getting-started).

>  * perhaps mention that `plt.show()` is sometimes necessary to actually show the plots, depending on the context (e.g. okay in Jupyter, but less so when copy-pasting each tut to a file).

Also a good point. I've added a statement about this as well at the same location as above.
