### 2024-09-15 #7134 @phibeck
> @phibeck Hi Sophie, sorry for being late but I just came back from my holidays and I found a _nice_ surprise, I have to move all my office elsewhere because of some maintenance work. Therefore if two reviewers are enough for you, please go ahead without my review.

_Originally posted by @sabinomaggi in https://github.com/openjournals/joss-reviews/issues/6991#issuecomment-2348595273_

Okay, thanks for letting me know. If we have two completed reviews that would be fine. In the event that another reviewer drops out, may I contact you again or rather not at the moment?
