### 2024-09-15 #6991 @phibeck
Thanks @sabinomaggi for your response. Since this Pre Review issue is closed, I copied your response into the Review issue and answered there.

> Okay, thanks for letting me know. If we have two completed reviews that would be fine. In the event that another reviewer drops out, may I contact you again or rather not at the moment?

_Originally posted by @phibeck in https://github.com/openjournals/joss-reviews/issues/7134#issuecomment-2351464691_
