### 2024-09-15 #7134 @sabinomaggi
Of course. My problem at the moment is that having to manage the deadlines I cannot miss while moving all my stuff elsewhere is a huge loss of time. I hope things will get better in a couple of weeks, but I don't want to delay the management of this paper because of my hassles.
