### 2024-09-15 jabirali/bodge#3 @mdavezac
potential regression in benchmarks.py
I'm trying to run the benchmarks for openjournals/joss-reviews#7134, but I'm getting:

```ipython
Traceback (most recent call last):
  File "./misc/benchmark.py", line 143, in <module>
    H1, T1 = bench_bodge(L, W)
  File "./misc/benchmark.py", line 120, in bench_bodge
    H = system(format="csr")
TypeError: 'Hamiltonian' object is not callable
```

Any ideas?
