### 2024-09-15 jabirali/bodge#3 @jabirali
Hi, thanks for checking the benchmarks.

I changed the `Hamiltonian` class a bit before submitting to JOSS, including that I renamed `Hamiltonian.__call__(format=...)` (which implemented the magic `system(format=...)` syntax above) to `Hamiltonian.matrix(format=...)` to make the conversion more explicit. I think the only change needed is to change `system(...)` to `system.matrix(...)` in the benchmark code, but I'll have a quick look now to see if I've overlooked anything else :)
