### 2024-09-15 jabirali/bodge#3 @jabirali
Thanks for notifying me about this.

I pushed a quick fix to the `main` branch now, with the update mentioned above the benchmark script is working again for me :).

Remember also to do a `pip install kwant` if you haven't already, to fetch the library that I'm benchmarking against as well. I added a quick note of this to the top of the benchmark script now.
