### 2024-09-15 a85a959
docs: updated footnote
