+++ tutorial.qmd
@@ -92,7 +92,7 @@ Some things I want to elaborate on in this example:
 - When filling out the matrix, we can use `lattice.sites()` and `lattice.bonds()` to iterate over the whole lattice. In this case, we have kept it simple, and only have one type of on-site term and one type of hopping term, but Bodge itself is very flexible. You can e.g. use `lattice.bonds(axis=0)` to iterate over only bonds that point along the 0th axis (x axis), which is useful if you need different hopping terms in different directions. Moreover, each coordinate $i$ above is simply a tuple $(x_i, y_i, z_i)$ where $0 \leq x_i < L_x$ and so on. Thus, you can use e.g. `if i[0] < Lx/2:` to implement a sharp interface between different materials, or a function call like `np.sin(np.pi*i[0]/Lx)` to implement a field that varies smoothly throughout the lattice.
 - The `system.ldos` method implements an efficient sparse matrix algorithm to obtain the local density of states at a single site $i$. The algorithm implemented is explained in detail in Appendix A of [this paper](https://doi.org/10.1103/PhysRevB.109.174506), and is essentially a simplified version of the algorithm from [this paper](https://doi.org/10.7566/jpsj.86.014708). You could alternatively diagonalize the Hamiltonian using `E, X = system.diagonalize()`, and then use the eigenvalues `E[n]` and corresponding eigenvectors `X[n,:,:]` to obtain the density of states as described in e.g. [Zhu's textbook](https://doi.org/10.1007/978-3-319-31314-6). However, this approach uses dense matrices (NumPy arrays), and is therefore significantly slower for large systems. I'd therefore recommend researching whether you can use a sparse matrix algorithm before you diagonalize large matrices.
 
-[^1]: Examples: Vim has built-in "digraphs" where you can press `<C-k>s*` in insert mode to type `σ`. Emacs has a built-in "TeX input method", which means that after pressing `C-\` you can type `\sigma` to insert the symbol `σ`. Most other editors code have third-party plugins for this, e.g. [Unicode Latex](https://marketplace.visualstudio.com/items?itemName=oijaz.unicode-latex) for VSCode. Another option is to use an OS-wide snippet expander (e.g. TextExpander), which you can setup to e.g. convert `;s` to `σ` in any application. There are many alternatives!
+[^1]: Examples: Vim has built-in "digraphs" where you can press `<C-k>s*` in insert mode to type `σ`. Emacs has a built-in "TeX input method", which means that after pressing `C-\` you can type `\sigma` to insert the symbol `σ`. Most other editors code have third-party plugins for this, e.g. [Unicode Latex](https://marketplace.visualstudio.com/items?itemName=oijaz.unicode-latex) for VSCode. Another option is to use an OS-wide snippet expander (e.g. TextExpander), which you can setup to e.g. convert `;s` to `a` in any application. Another alternative is to just enable a Greek keyboard layout in your OS settings, with a hotkey to switch between the layouts. There are many alternatives!
 
 ## Conventional superconductor
 
