### 2024-09-19 jabirali/bodge#4 @yw-fang
[REVIEW comments] for the paper and code submitted to JOSS
For the submitted work https://github.com/openjournals/joss-reviews/issues/7134, I found this code Bodge was and is still being developed by @jabirali with substantial scholarly effort and all the basic features of a scientific code. This code can model many materials systems including the conventional/unconventional superconductors and magnetic systems once the Hamiltonian was properly given.  The code is well written and the examples are easy to follow. I think this work basically meet the criteria for the publication at JOSS.

Herein, I list several comments on the manuscript and code itself that you may consider addressing before the accept for publication.

## Manuscript

- [x] While I am also working on the superconductivity (mostly on prediction of high Tc superconducting hydrides), it's dramatically different from the research direction you've moving forward. So I am not an expert in your field.
My question on the manuscript is that Bodge was said to be able to study the superconductivity in magnetic heterostructures, could you list several more details properties (e.g. Tc) of superconductivity if Bodge can address in the manuscript? It is not very clear to me what superconducting properties it can study in the present version.

- [x] There is one more minor comment on the manuscript. Although the CSR has been popular as a format, I think it's better to give the full definition in the main text for some readers who are not familiar with it.

## Code
- [x] The test suite needs to be further improved so that the core functions can be tested more thoroughly. The unittest commands are recommended to be shown in README or the documentation.
- [x] You may consider adding the URL of the doc to the upper right side of the code page in Github so that the users can easily find the documentation.
