### 2024-09-19 5ef32a9
docs: add badges to readme
