### 2024-09-19 fffc556
add: mention tests in readme
