### 2024-09-20 jabirali/bodge#3 @jabirali
I guess the benchmark script is fixed now, so I'll close this issue as resolved. But please feel free to reopen it if you have any more problems, and I'd be happy to help.
