### 2024-09-21 #7134 @mdavezac
The author has addressed all the issues I have raised. I am happy to recommend Bodge for publication in JOSS!
