### 2024-09-22 325e5b3
test: add tests of matrix format types
