### 2024-09-23 #7134 @phibeck
Hi @mdavezac and @yw-fang, thank you for your thorough reviews! :tada: @yw-fang I see that the Automated tests box on your checklist isn't ticked off yet. If your questions have been addressed adequately, please check off the box when you have a moment, thanks.

@jabirali thank you for addressing the reviewers' comments and for the update. From your summary in https://github.com/jabirali/bodge/issues/4 I see two tasks for the manuscript and one for the test suite. Do you intend to include any changes to the test suite before we move on here, or is this future development? I will wait for any changes to the manuscript before we proceed.
