### 2024-09-23 1d900dd
test: improvement and documentation of existing test suite
