+++ tests/test_hamiltonian.py
@@ -1,4 +1,12 @@
+"""Unit tests for the `Hamiltonian` class and some of its core methods.
+
+In this set of tests, we mainly try to verify that the implementation
+is mathematically correct. For more physically minded tests, please
+see the integration tests in `test_physics.py`.
+"""
+
 import numpy as np
+from numpy.random import random as r
 from pytest import raises
 from scipy.linalg import eigh
 
@@ -7,6 +15,15 @@ from bodge.common import *
 
 
 def test_matrix_hermitian():
+    """Test that a semi-random Hamiltonian matrix becomes Hermitian.
+
+    Note that for the constructed Hamiltonian to be Hermitian, we
+    need (i) a symmetry between hopping i->j and j->i, and (ii)
+    each submatrix we define below to itself be Hermitian. Within
+    these bounds, we add some randomness for more general testing.
+
+    For non-Hermitian matrices, we test that an exception is raised.
+    """
     # Instantiate a somewhat dense complex Hamiltonian. Note that
     # the submatrices need to be Hermitian for the whole to be so.
     lattice = CubicLattice((3, 5, 7))
@@ -14,29 +31,40 @@ def test_matrix_hermitian():
 
     with system as (H, Δ):
         for i in lattice.sites():
-            H[i, i] = 1 * σ3 + 2 * σ2
-            Δ[i, i] = 5 * σ0 - 3 * σ2
+            H[i, i] = r() * σ0 + r() * σ1 + r() * σ2 + r() * σ3
+            Δ[i, i] = (r() * σ1 + r() * σ2 + r() * σ3) @ jσ2
 
         for i, j in lattice.bonds():
-            H[i, j] = 3 * σ0 - 4 * σ2
-            Δ[i, j] = 2 * σ3 + 5 * σ2
+            t = [r(), r(), r(), r()]
+            H[i, j] = t[0] * σ0 + t[1] * σ1 + t[2] * σ2 + t[3] * σ3
+            H[j, i] = t[0] * σ0 + t[1] * σ1 + t[2] * σ2 + t[3] * σ3
+            Δ[i, j] = (r() * σ1 + r() * σ2 + r() * σ3) @ jσ2
 
         for i, j in lattice.edges():
-            H[i, j] = 7 * σ0 - 13 * σ2
-            Δ[i, j] = 9 * σ3 + 11 * σ2
+            T = [r(), r(), r(), r()]
+            H[i, j] = T[0] * σ0 + T[1] * σ1 + T[2] * σ2 + T[3] * σ3
+            H[j, i] = T[0] * σ0 + T[1] * σ1 + T[2] * σ2 + T[3] * σ3
+            Δ[i, j] = (r() * σ1 + r() * σ2 + r() * σ3) @ jσ2
 
     # Verify that the result is Hermitian.
     H = system._matrix.todense()
     assert np.allclose(H, H.T.conj())
 
     # Check non-hermiticity warnings.
-    with raises(RuntimeError):
+    with raises(Exception):
         i = (1, 1, 1)
         with system as (H, Δ):
             H[i, i] = 1j * σ1
 
 
 def test_matrix_export():
+    """Test that the Hamiltonian matrix is correctly exported.
+
+    This entails that the same matrix should be constructed both when
+    exporting to sparse and dense formats, and that an exception
+    should be raised if an unimplemented format is requested.
+    """
+
     # Instantiate a somewhat random test system.
     lattice = CubicLattice((3, 5, 7))
     system = Hamiltonian(lattice)
@@ -54,9 +82,9 @@ def test_matrix_export():
 
     # Check that each matrix has the right format.
[…]
