### 2024-09-23 26f4e4d
test: added several new integration tests
