+++ tests/test_physics.py
@@ -1,0 +1,294 @@
+"""Integration tests that validate some known physical phenomena using Bodge.
+
+In contrast to `test_lattice.py` and `test_hamiltonian.py`, these are
+not intended to test the mathematical properties of individual classes
+or methods, but rather to combine features from across Bodge in order
+to verify that important physical phenomena are modeled correctly.
+"""
+
+import numpy as np
+from numpy.random import random as r
+
+from bodge import *
+from bodge.common import *
+
+
+def test_superconducting_gap_existence():
+    """When a normal metal becomes superconducting a "gap" opens.
+
+    There's two ways to test this:
+
+    1. The local density of states disappears within the gap;
+    2. The lowest energy eigenvalue is pushed above the gap.
+
+    Here, we verify this result using both approaches.
+    """
+    # Instantiate a normal metal.
+    lattice = CubicLattice((16, 16, 1))
+    system = Hamiltonian(lattice)
+
+    with system as (H, Δ):
+        for i in lattice.sites():
+            H[i, i] = -1.5 * σ0
+        for i, j in lattice.bonds():
+            H[i, j] = -1.0 * σ0
+
+    # Calculate the central LDOS.
+    Δs = 0.5
+    i = (8, 8, 0)
+    ω = np.array([-1.2 * Δs, -0.8 * Δs, +0.8 * Δs, 1.2 * Δs])
+    ρ1 = system.ldos(i, ω)
+
+    # Calculate the lowest eigenvalue.
+    ε, _ = system.diagonalize()
+    ε1 = np.min(ε)
+
+    # Add superconductivity.
+    with system as (H, Δ):
+        for i in lattice.sites():
+            Δ[i, i] = Δs * jσ2
+
+    # Calculate the central LDOS.
+    ρ2 = system.ldos(i, ω)
+
+    # Calculate the lowest eigenvalue.
+    ε, _ = system.diagonalize()
+    ε2 = np.min(ε)
+
+    # The LDOS should decrease inside the gap of a superconductor.
+    assert ρ2[1] < ρ1[1]
+    assert ρ2[2] < ρ1[2]
+
+    # The LDOS should increase outside the gap of a superconductor.
+    assert ρ2[0] > ρ1[0]
+    assert ρ2[3] > ρ1[3]
+
+    # The lowest energy eigenvalue should increase in a superconductor.
+    assert ε2 > ε1
+
+def test_superconducting_gap_scaling():
+    """The superconducting energy gap is given by its order parameter.
+
+    We can test this by instantiating superconductors with varying
+    order parameters, and then use the system's lowest energy
+    eigenvalue as a measure of the energy gap in the spectrum.
+    Ideally, these should be identical, but due to e.g. finite size
+    effects in quick tests there may be small variations. We therefore
+    settle for testing that one increases monotonically with the other.
+    """
+    # Construct a simple lattice.
+    lattice = CubicLattice((32, 1, 1))
+    system = Hamiltonian(lattice)
+
+    # Define some representative order parameters.
+    Δ_in = [0.0, 0.01, 0.03, 0.1, 0.3, 1.0]
+    Δ_out = []
+
+    # Construct the base Hamiltonian.
+    t = 1.0
+    μ = 1.5 * t
+    with system as (H, Δ):
+        for i in lattice.sites():
+            H[i, i] = -μ * σ0
+        for i, j in lattice.bonds():
+            H[i, j] = -t * σ0
+
+    # Perform the calculations.
+    for Δ0 in Δ_in:
+        # Update the order parameter.
+        with system as (H, Δ):
[…]
