### 2024-09-23 d6abe41
test: added proximity effect test
