### 2024-09-23 jabirali/bodge#4 @jabirali
@yw-fang:

From before, I had a set of unit tests that provided near-100% test coverage, but focused more on "mathematical" tests: checking that symmetries are correct, matrices are Hermitian, that correct exceptions are raised for invalid input, and so on.

I have now extended and documented the existing test suite to reach exactly 100% test coverage, and have in addition **added a file `test_physics.py` which includes 8 new physics-based "integration tests".** These new tests validate different combinations of features from Bodge to see that various numerical experiments on superconductors and magnets yield known physical predictions expected from theory.

I will try to revise the manuscript as discussed later this week.
