### 2024-09-24 9cdc9ad
paper: added discussion of sparse matrices
