+++ paper.bib
@@ -92,3 +92,15 @@
 	url = {https://doi.org/10.5281/zenodo.4010216}
 }
 
+@article{nagai2020a,
+  title = {N-Independent Localized {{Krylov}}–{{Bogoliubov-de Gennes}} Method: {{Ultra-fast}} Numerical Approach to Large-Scale Inhomogeneous Superconductors},
+  author = {Nagai, Yuki},
+  date = {2020},
+  journaltitle = {Journal of the Physical Society of Japan},
+  shortjournal = {J Phys Soc Jpn},
+  volume = {89},
+  number = {7},
+  pages = {074703},
+  issn = {0031-9015},
+  doi = {10.7566/jpsj.89.074703},
+}
