### 2024-09-24 5f5ca3e
paper: added sparse matrix algorithm references
