+++ paper.bib
@@ -1,57 +1,10 @@
-@article{ouassou_alt_2023,
-	title = {{DC} {Josephson} effect in altermagnets},
-	volume = {131},
-	issn = {0031-9007, 1079-7114},
-	url = {https://link.aps.org/doi/10.1103/PhysRevLett.131.076003},
-	doi = {10.1103/PhysRevLett.131.076003},
-	language = {en},
-	number = {7},
-	urldate = {2023-08-18},
-	journal = {Physical Review Letters},
-	author = {Ouassou, Jabir Ali and Brataas, Arne and Linder, Jacob},
-	month = aug,
-	year = {2023},
-	pages = {076003},
-}
-
-@article{ouassou_rkky_2024,
-	title = {{RKKY} interaction in triplet superconductors: {Dzyaloshinskii}{Moriya}--type interaction mediated by spin-polarized {Cooper} pairs},
-	volume = {109},
-	issn = {2469-9950, 2469-9969},
-	shorttitle = {{RKKY} interaction in triplet superconductors},
-	url = {https://link.aps.org/doi/10.1103/PhysRevB.109.174506},
-	doi = {10.1103/PhysRevB.109.174506},
-	language = {en},
-	number = {17},
-	urldate = {2024-05-03},
-	journal = {Physical Review B},
-	author = {Ouassou, Jabir Ali and Yokoyama, Takehito and Linder, Jacob},
-	month = may,
-	year = {2024},
-	pages = {174506},
-}
-
-@article{ouassou_dmi_2024,
-	title = {{Dzyaloshinskii}--{Moriya} spin--spin interaction from mixed-parity superconductivity},
-	url = {http://arxiv.org/abs/2407.07144},
-	doi = {10.48550/arXiv.2407.07144},
-	urldate = {2024-07-11},
-	publisher = {arXiv},
-	author = {Ouassou, Jabir Ali and Yokoyama, Takehito and Linder, Jacob},
-	month = jul,
-	year = {2024},
-	note = {arXiv:2407.07144 [cond-mat]},
-}
-
-@article{scipy_2020,
-	author = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and {van der Walt}, St{\'e}fan J. and Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and Kern, Robert and Larson, Eric and Carey, C J and Polat, {\.I}lhan and Feng, Yu and Moore, Eric W. and {VanderPlas}, Jake and Laxalde, Denis and Perktold, Josef and Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and Harris, Charles R. and Archibald, Anne M. and Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
-	title = {{{SciPy} 1.0: Fundamental Algorithms for Scientific Computing in Python}},
-	journal = {Nature Methods},
-	year = {2020},
-	volume = {17},
-	pages = {261--272},
-	adsurl = {https://rdcu.be/b08Wh},
-	doi = {10.1038/s41592-019-0686-2},
+@phdthesis{benfenati_fermi_2022,
+  title = {Numerical Solutions to Non-Linear Inhomogeneous Problems in Superconductivity},
+  author = {Benfenati, Andrea Ludovico},
+  year = {2022},
[…]
