### 2024-09-24 342aafa
paper: added reference to fermi matrix
