+++ paper.md
@@ -49,7 +49,7 @@ There are two main alternatives that arguably fill a similar niche to Bodge: Kwa
 
 # Examples
 
-Introductory examples of how to use Bodge are provided in the [official documentation](https://jabirali.github.io/bodge/). Examples of research problems that have been studied using Bodge include superconductor/altermagnet heterostructures [@ouassou_alt_2023] and RKKY interactions in unconventional superconductors [@ouassou_rkky_2024; @ouassou_dmi_2024]. Two of those papers describe sparse matrix algorithms that have been used to speed up numerical calculations based on Bodge [@ouassou_alt_2023; @ouassou_rkky_2024], including Chebyshev expansion of the Fermi matrix [@weisse_chebyshev_2006; @benfenati_fermi_2022] and a quick way to calculate the local density of states [@nagai_ldos_2017].
+Introductory examples of how to use Bodge are provided in the [official documentation](https://jabirali.github.io/bodge/). Examples of research problems that have been studied using Bodge include superconductor/altermagnet heterostructures [@ouassou_alt_2023] and RKKY interactions in unconventional superconductors [@ouassou_rkky_2024; @ouassou_dmi_2024]. These papers also describe some sparse matrix algorithms that have been used to speed up numerical calculations based on Bodge, including Chebyshev expansion of the Fermi matrix [@weisse_chebyshev_2006; @benfenati_fermi_2022; @goedecker_fermi_1994] and a quick way to calculate the local density of states [@nagai_ldos_2017].
 
 For a simple example of how this package can be used, consider a $64a\times64a$ square lattice. Let's assume that the whole system is a metal with chemical potential $\mu = 1.5t$, where $t=1$ is the hopping amplitude. Moreover, let's assume that half the system ($x < 32a$) is a conventional $s$-wave superconductor with an order parameter $\Delta_s = 0.1t$, whereas the other half is a ferromagnet with exchange field $\mathbf{M} = M_z \mathbf{e}_z$. To construct the Hamiltonian matrix for such a system in the CSR sparse format, we can use the following code:
 ```python
