+++ paper.bib
@@ -5,6 +5,7 @@
   address = {Stockholm, Sweden},
   langid = {english},
   school = {KTH},
+  url = {https://urn.kb.se/resolve?urn=urn:nbn:se:kth:diva-311403},
 }
 
 @book{degennes_bdg_1966,
@@ -12,6 +13,19 @@
 	isbn = {978-0-7382-0101-6},
 	author = {{de Gennes}, Pierre Gilles},
 	year = {1966},
+}
+
+
+@article{goedecker_fermi_1994,
+  title = {Efficient Linear Scaling Algorithm for Tight-Binding Molecular Dynamics},
+  author = {Goedecker, S. and Colombo, L.},
+  year = {1994},
+  journal = {Physical Review Letters},
+  volume = {73},
+  number = {1},
+  pages = {122--125},
+  issn = {0031-9007},
+  doi = {10.1103/physrevlett.73.122},
 }
 
 @article{groth_kwant_2014,
