### 2024-09-24 fb92e60
fix: add missing type checks and hints
