### 2024-09-24 a2ec269
paper: added more info about workflows
