+++ paper.md
@@ -47,7 +47,7 @@ There are two main alternatives that arguably fill a similar niche to Bodge: Kwa
 
 *Sparse matrix formats*, such as e.g. the *Compressed Sparse Row* (CSR) format used in the example below, have the advantage that they basically only store the non-zero elements of a matrix. In the case of a typical tight-binding model with nearest-neighbor hopping terms, the Hamiltonian matrix of a system of $N$ atoms has $\mathcal{O}(N^2)$ elements where only $\mathcal{O}(N)$ are actually non-zero. Thus, algorithms that leverage sparse matrices often result in at least an $\mathcal{O}(N)$ numerical speed-up, which becomes highly significant for large systems. With additional approximations, even larger speed-ups can in some cases be possible [@nagai_krylov_2020], although it depends on your system size and the strength of its interactions whether such approximations provide a good trade-off between accuracy and speed. However, *dense matrices* (i.e. NumPy arrays) allow for simpler solution algorithms based on e.g. full matrix diagonalization, and can become faster than sparse matrix algorithms for small systems or when leveraging GPU calculations. For this reason, both sparse and dense matrices are fully supported by Bodge, allowing the user to pick the most suitable matrix format for the task at hand.
 
-# Examples
+# Examples and workflows
 
 Introductory examples of how to use Bodge are provided in the [official documentation](https://jabirali.github.io/bodge/). Examples of research problems that have been studied using Bodge include superconductor/altermagnet heterostructures [@ouassou_alt_2023] and RKKY interactions in unconventional superconductors [@ouassou_rkky_2024; @ouassou_dmi_2024]. These papers also describe some sparse matrix algorithms that have been used to speed up numerical calculations based on Bodge, including Chebyshev expansion of the Fermi matrix [@weisse_chebyshev_2006; @benfenati_fermi_2022; @goedecker_fermi_1994] and a quick way to calculate the local density of states [@nagai_ldos_2017].
 
@@ -74,10 +74,24 @@ with system as (H, Δ):
             Δ[i, i] = -Δs * jσ2
         else:
             H[i, i] = -μ * σ0 - Mz * σ3
+```
+Note the use of a context manager (`with`-block) to provide an intuitive array syntax for accessing the relevant parts of the Hamiltonian matrix, while abstracting away the underlying sparse matrix details. Afterwards, there are several different ways to use the resulting `Hamiltonian` object.
 
+Some physical observables can be directly calculated using the methods provided in Bodge. For instance, one can use the method `system.ldos` to directly calculate the local density of states, which can then be used to check for spectral features such as e.g. a *superconducting gap* or a *zero-energy peak*. There is also a method `system.free_energy` which calculates the *free energy* of the system. By varying parameters in the Hamiltonian (e.g. the orientation of a magnetic field) and then minimizing this free energy, one can e.g. determine the ground state of the system.
+
+Most calculations of interest, however, requires that the user implements some code themselves. There are then two main approaches one can take. The classic approach is *matrix diagonalization* which uses dense matrices internally.[^1] Bodge provides the method `diagonalize` for this purpose:
+```python
+E, X = system.diagonalize()
+```
+The results above contain the positive energies $E_n$ and corresponding  $\mathbf{X}_n$ which satisfy the eigenvalue equation $\mathbf{H} \mathbf{X}_n = E_n \mathbf{X}_n$ [^2]. Once the eigenvalues and eigenvectors have been obtained, the user can themselves calculate physical properties of interest from on these using equations from standard textbooks on the "Bogoliubov–de Gennes" approach to modeling superconductivity [@zhu_bdg_2016]. It is a future goal to incorporate more calculation methods of this kind into the Bodge package itself.
+
[…]
