+++ paper.bib
@@ -8,13 +8,24 @@
   url = {https://urn.kb.se/resolve?urn=urn:nbn:se:kth:diva-311403},
 }
 
-@book{degennes_bdg_1966,
-	title = {Superconductivity of metals and alloys},
-	isbn = {978-0-7382-0101-6},
-	author = {{de Gennes}, Pierre Gilles},
-	year = {1966},
+@article{covaci_cbdg_2010,
+  title = {Efficient Numerical Approach to Inhomogeneous Superconductivity: The Chebyshev-Bogoliubov--de Gennes Method},
+  author = {Covaci, L. and Peeters, F. M. and Berciu, M.},
+  year = {2010},
+  journal = {Physical Review Letters},
+  volume = {105},
+  number = {16},
+  pages = {167006},
+  issn = {0031-9007},
+  doi = {10.1103/physrevlett.105.167006},
 }
 
+@book{degennes_bdg_1966,
+  title = {Superconductivity of metals and alloys},
+  isbn = {978-0-7382-0101-6},
+  author = {{de Gennes}, Pierre Gilles},
+  year = {1966},
+}
 
 @article{goedecker_fermi_1994,
   title = {Efficient Linear Scaling Algorithm for Tight-Binding Molecular Dynamics},
@@ -29,26 +40,26 @@
 }
 
 @article{groth_kwant_2014,
-	doi = {10.1088/1367-2630/16/6/063065},
-	url = {https://dx.doi.org/10.1088/1367-2630/16/6/063065},
-	year = {2014},
-	month = {jun},
-	publisher = {IOP Publishing},
-	volume = {16},
-	number = {6},
-	pages = {063065},
-	author = {Christoph W Groth and Michael Wimmer and Anton R Akhmerov and Xavier Waintal},
-	title = {Kwant: a software package for quantum transport},
-	journal = {New Journal of Physics},
+  doi = {10.1088/1367-2630/16/6/063065},
+  url = {https://dx.doi.org/10.1088/1367-2630/16/6/063065},
+  year = {2014},
+  month = {jun},
+  publisher = {IOP Publishing},
+  volume = {16},
+  number = {6},
+  pages = {063065},
+  author = {Christoph W Groth and Michael Wimmer and Anton R Akhmerov and Xavier Waintal},
+  title = {Kwant: a software package for quantum transport},
+  journal = {New Journal of Physics},
 }
 
 @article{moldovan_pybinding_2020,
-	author = {Moldovan, Dean, and Anđelković, Miša, and Peeters, Francois},
-	year = {2020},
-	title ={pybinding: a Python package for tight-binding calculations},
-	journal = {Zenodo},
-	doi = {10.5281/zenodo.4010216},
-	url = {https://doi.org/10.5281/zenodo.4010216}
+  author = {Moldovan, Dean, and Anđelković, Miša, and Peeters, Francois},
+  year = {2020},
+  title ={pybinding: a Python package for tight-binding calculations},
+  journal = {Zenodo},
+  doi = {10.5281/zenodo.4010216},
+  url = {https://doi.org/10.5281/zenodo.4010216}
 }
 
 @article{nagai_krylov_2020,
@@ -77,59 +88,59 @@
 }
 
 @article{ouassou_alt_2023,
-	title = {{DC} {Josephson} effect in altermagnets},
-	volume = {131},
-	issn = {0031-9007, 1079-7114},
-	url = {https://link.aps.org/doi/10.1103/PhysRevLett.131.076003},
-	doi = {10.1103/PhysRevLett.131.076003},
-	language = {en},
-	number = {7},
-	urldate = {2023-08-18},
[…]
