### 2024-09-24 08bf506
fix: footnotes
