### 2024-09-24 be48de1
fix: bibliography
