+++ paper.bib
@@ -17,7 +17,7 @@
   number = {16},
   pages = {167006},
   issn = {0031-9007},
-  doi = {10.1103/physrevlett.105.167006},
+  doi = {10.1103/PhysRevLett.105.167006},
 }
 
 @book{degennes_bdg_1966,
@@ -36,7 +36,7 @@
   number = {1},
   pages = {122--125},
   issn = {0031-9007},
-  doi = {10.1103/physrevlett.73.122},
+  doi = {10.1103/PhysRevLett.73.122},
 }
 
 @article{groth_kwant_2014,
@@ -139,8 +139,7 @@
   year = {2020},
   volume = {17},
   pages = {261--272},
-  adsurl = {https://rdcu.be/b08Wh},
-  doi = {10.1038/s41592-019-0686-2},
+  doi = {10.1038/S41592-019-0686-2},
 }
 
 @article{weisse_chebyshev_2006,
@@ -151,7 +150,7 @@
   volume = {78},
   number = {1},
   pages = {275--306},
-  doi = {10.1103/revmodphys.78.275},
+  doi = {10.1103/RevModPhys.78.275},
  }
 
 @book{zhu_bdg_2016,
