### 2024-09-24 226b90f
paper: fix formatting
