### 2024-09-24 f669f96
paper: revise eigenvalue notation
