+++ paper.md
@@ -81,9 +81,9 @@ Some physical observables can be directly calculated using the methods provided 
 
 Most calculations of interest, however, requires that the user implements some code themselves. There are then two main approaches one can take. The classic approach is *matrix diagonalization* which uses dense matrices internally. Bodge provides the method `diagonalize` for this purpose:
 ```python
-E, X = system.diagonalize()
+ε, χ = system.diagonalize()
 ```
-The results above contain the positive energies $E_n$ and corresponding  $\mathbf{X}_n$ which satisfy the eigenvalue equation $\mathbf{H} \mathbf{X}_n = E_n \mathbf{X}_n$. (Only positive eigenvalues are returned due to the so-called "Nambu doubling" of degrees of freedom in superconducting systems.) Once the eigenvalues and eigenvectors have been obtained, the user can themselves calculate physical properties of interest from on these using equations from standard textbooks on the "Bogoliubov–de Gennes" approach to modeling superconductivity [@zhu_bdg_2016]. It is a future goal to incorporate more calculation methods of this kind into the Bodge package itself. Moreover, support for matrix diagonalization using GPUs (via the CuPy package) is currently under development since it offers orders of magnitude faster computation when the right hardware is available.
+The results above contain the positive energies $\varepsilon_n$ and corresponding vectors $\boldsymbol{\chi}_n$ which satisfy the eigenvalue equation $\mathbf{H} \boldsymbol{\chi}_n = \varepsilon_n \boldsymbol{\chi}_n$. (Only positive eigenvalues are returned due to the so-called "Nambu doubling" of degrees of freedom in superconducting systems.) Once the eigenvalues and eigenvectors have been obtained, the user can themselves calculate physical properties of interest from on these using equations from standard textbooks on the "Bogoliubov–de Gennes" approach to modeling superconductivity [@zhu_bdg_2016]. It is a future goal to incorporate more calculation methods of this kind into the Bodge package itself. Moreover, support for matrix diagonalization using GPUs (via the CuPy package) is currently under development since it offers orders of magnitude faster computation when the right hardware is available.
 
 Examples of physical observables that can be calculated from the eigenvalues and eigenvectors include e.g. the superconducting order parameter, electric currents, and spin currents. Indirectly, these can in turn be used to calculate even more physical observables. For instance, the critical current of a Josephson junction can be defined as the largest electric current that can flow through it for any phase difference, and the critical temperature of a bulk superconductor can be defined as the largest temperature at which the superconducting order parameter remains non-zero.
 
