### 2024-09-24 c3068d2
paper: fix references
