+++ paper.md
@@ -49,7 +49,7 @@ There are two main alternatives that arguably fill a similar niche to Bodge: Kwa
 
 # Examples and workflows
 
-Introductory examples of how to use Bodge are provided in the [official documentation](https://jabirali.github.io/bodge/). Examples of research problems that have been studied using Bodge include superconductor/altermagnet heterostructures [@ouassou_alt_2023] and RKKY interactions in unconventional superconductors [@ouassou_rkky_2024; @ouassou_dmi_2024]. These papers also describe some sparse matrix algorithms that have been used to speed up numerical calculations based on Bodge, including Chebyshev expansion of the Fermi matrix [@ouassou_alt_23; @weisse_chebyshev_2006; @benfenati_fermi_2022; @goedecker_fermi_1994] and a quick way to calculate the local density of states [@ouassou_rkky_24; @nagai_ldos_2017].
+Introductory examples of how to use Bodge are provided in the [official documentation](https://jabirali.github.io/bodge/). Examples of research problems that have been studied using Bodge include superconductor/altermagnet heterostructures [@ouassou_alt_2023] and RKKY interactions in unconventional superconductors [@ouassou_rkky_2024; @ouassou_dmi_2024]. These papers also describe some sparse matrix algorithms that have been used to speed up numerical calculations based on Bodge, including Chebyshev expansion of the Fermi matrix [@ouassou_alt_2023; @weisse_chebyshev_2006; @benfenati_fermi_2022; @goedecker_fermi_1994] and a quick way to calculate the local density of states [@ouassou_rkky_2024; @nagai_ldos_2017].
 
 For a simple example of how this package can be used, consider a $64a\times64a$ square lattice. Let's assume that the whole system is a metal with chemical potential $\mu = 1.5t$, where $t=1$ is the hopping amplitude. Moreover, let's assume that half the system ($x < 32a$) is a conventional $s$-wave superconductor with an order parameter $\Delta_s = 0.1t$, whereas the other half is a ferromagnet with exchange field $\mathbf{M} = M_z \mathbf{e}_z$. To construct the Hamiltonian matrix for such a system in the CSR sparse format, we can use the following code:
 ```python
@@ -83,7 +83,7 @@ Most calculations of interest, however, requires that the user implements some c
 ```python
 ε, χ = system.diagonalize()
 ```
-The results above contain the positive energies $\varepsilon_n$ and corresponding vectors $\boldsymbol{\chi}_n$ which satisfy the eigenvalue equation $\mathbf{H} \boldsymbol{\chi}_n = \varepsilon_n \boldsymbol{\chi}_n$. (Only positive eigenvalues are returned due to the so-called "Nambu doubling" of degrees of freedom in superconducting systems.) Once the eigenvalues and eigenvectors have been obtained, the user can themselves calculate physical properties of interest from on these using equations from standard textbooks on the "Bogoliubov–de Gennes" approach to modeling superconductivity [@zhu_bdg_2016]. It is a future goal to incorporate more calculation methods of this kind into the Bodge package itself. Moreover, support for matrix diagonalization using GPUs (via the CuPy package) is currently under development since it offers orders of magnitude faster computation when the right hardware is available.
[…]
