### 2024-09-25 bd18298
paper: improve text
