+++ paper.md
@@ -45,7 +45,7 @@ To do anything useful with that Hamiltonian *on a computer*, however, you typica
 
 There are two main alternatives that arguably fill a similar niche to Bodge: Kwant [@groth_kwant_2014] and Pybinding [@moldovan_pybinding_2020]. Compared to these packages, the main benefit of Bodge is the focus on the BdG Hamiltonian in particular. For instance, using Kwant, it is up to the user to declare that each lattice site has 4 degrees of freedom (spin-up electrons, spin-down electrons, spin-up holes, and spin-down holes), and to ensure that you construct a Hamiltonian with the correct particle-hole symmetries. Bodge, however, *assumes* that these are the only relevant degrees of freedom, and enforces the relevant symmetries by default. In practice, this means that Kwant can be used to study a broader variety of physical systems, whereas Bodge can provide a friendlier syntax for users who work specifically on superconducting systems. Both packages support both NumPy arrays and SciPy sparse matrices as output formats, and both provide similar performance in the limit of large systems.
 
-Sparse matrices, including e.g. the *Compressed Sparse Row* (CSR) format used below, have the advantage that they only store the non-zero elements of a matrix. For a typical tight-binding model with nearest-neighbor hopping terms, the Hamiltonian matrix that describes a lattice with $N$ atoms has $\mathcal{O}(N^2)$ elements where only $\mathcal{O}(N)$ are non-zero. Thus, algorithms that leverage sparse matrices often result in at least an $\mathcal{O}(N)$ reduction in CPU and RAM requirements, which becomes highly significant for large systems ($>10^5$ atoms). With additional approximations, even larger improvements can in some cases be possible [@nagai_krylov_2020], although it depends on your system size and the strength of its interactions whether such approximations provide a good trade-off between accuracy and speed. However, *dense matrices* (i.e. NumPy arrays) allow for simpler solution algorithms based on e.g. full matrix diagonalization, and can become faster than sparse matrix algorithms for small systems or when leveraging GPU acceleration. For this reason, both sparse and dense matrices are fully supported by Bodge, allowing the user to pick the most suitable matrix format for the task at hand.
+Sparse matrices, including e.g. the *Compressed Sparse Row* (CSR) format used below, have the advantage that they only store the non-zero elements of a matrix. For a typical tight-binding model with nearest-neighbor hopping terms, the Hamiltonian matrix that describes a lattice with $N$ atoms has $\mathcal{O}(N^2)$ elements where only $\mathcal{O}(N)$ are non-zero. Thus, algorithms that leverage sparse matrices often result in at least an $\mathcal{O}(N)$ reduction in CPU and RAM requirements, which becomes highly significant for large systems. With additional approximations, even larger improvements can in some cases be possible [@nagai_krylov_2020], although it depends on your system size and the strength of its interactions whether such approximations provide a good trade-off between accuracy and speed. However, *dense matrices* (i.e. NumPy arrays) allow for simpler solution algorithms based on e.g. full matrix diagonalization, and can become faster than sparse matrix algorithms for small systems or when leveraging GPU acceleration. For this reason, both sparse and dense matrices are fully supported by Bodge, allowing the user to pick the most suitable matrix format for the task at hand.
 
 # Examples and workflows
 
