### 2024-09-25 dbe4b5e
paper: improve text
