### 2024-09-25 24e0737
paper: improve text
