### 2024-09-25 41738b1
paper: improve text
