+++ paper.md
@@ -51,7 +51,7 @@ Sparse matrices, including e.g. the *Compressed Sparse Row* (CSR) format used be
 
 Introductory examples of how to use Bodge are provided in the [official documentation](https://jabirali.github.io/bodge/). Examples of research problems that have been studied using Bodge include superconductor/altermagnet heterostructures [@ouassou_alt_2023] and RKKY interactions in unconventional superconductors [@ouassou_rkky_2024; @ouassou_dmi_2024]. These papers also describe some sparse matrix algorithms that can be used together with  Bodge, including Chebyshev expansion of the Fermi matrix [@ouassou_alt_2023; @weisse_chebyshev_2006; @benfenati_fermi_2022; @goedecker_fermi_1994] and a quick way to calculate the local density of states [@ouassou_rkky_2024; @nagai_ldos_2017].
 
-For a simple example of how this package can be used, consider a $64a\times64a$ square lattice. Let's assume that the whole system is a metal with chemical potential $\mu = 1.5t$, where $t=1$ is the hopping amplitude. Moreover, let's assume that half the system ($x < 32a$) is a conventional $s$-wave superconductor with an order parameter $\Delta_s = 0.1t$, whereas the other half is a ferromagnet with exchange field $\mathbf{M} = M_z \mathbf{e}_z$. To construct the Hamiltonian matrix for such a system in the CSR sparse format, we can use the following code:
+For a simple example of how this package can be used, consider a $64a\times64a$ square lattice. Let's assume that the whole system is a metal with chemical potential $\mu = 1.5t$, where $t=1$ is the hopping amplitude. Moreover, let's assume that half the system ($x < 32a$) is a conventional $s$-wave superconductor with an order parameter $\Delta_s = 0.1t$, whereas the other half is a ferromagnet with exchange field $\mathbf{M} = M_z \mathbf{e}_z$. To describe the Hamiltonian corresponding to this system we can use the following code:
 ```python
 from bodge import *
 
@@ -87,7 +87,7 @@ The results above contain the positive energies $E_n$ and corresponding vectors 
 
 Examples of physical observables that can be calculated from the eigenvalues and eigenvectors include e.g. the superconducting order parameter, electric currents, and spin currents. These can in turn be used to calculate even more material properties. For instance, the critical current of a Josephson junction is defined as the largest electric current that can flow through it for any phase difference, and the critical temperature of a bulk superconductor is defined as the largest temperature at which the superconducting order parameter remains non-zero.
 
-A modern alternative to matrix diagonalization is a series of algorithms based on Chebyshev expansion of the Hamiltonian matrix [@ouassou_alt_2023; @benfenati_fermi_2022; @covaci_cbdg_2010; @nagai_krylov_2020; @weisse_chebyshev_2006]. These algorithms take advantage of the extreme sparsity of the Hamiltonian matrix, and thus typically provide a performance benefit for very large lattices. One of the main design goals behind Bodge has been to make it trivial to construct sparse representations of the Hamiltonian matrix for this purpose, and it is therefore straight-forward to export the result as e.g. a CSR matrix:
+A modern alternative to matrix diagonalization is a series of algorithms based on Chebyshev expansion of the Hamiltonian matrix [@ouassou_alt_2023; @benfenati_fermi_2022; @covaci_cbdg_2010; @nagai_krylov_2020; @weisse_chebyshev_2006]. These algorithms take advantage of the extreme sparsity of the Hamiltonian matrix, and thus typically provide a performance benefit for very large lattices. One of the main design goals behind Bodge has been to make it trivial to construct sparse representations of the Hamiltonian matrix for this purpose, and it is therefore straight-forward to export the result as e.g. a CSR sparse matrix:
 ```python
 H = system.matrix(format="csr")
 ```
