### 2024-09-25 359b84f
paper: improve text
