### 2024-09-25 a8ec918
paper: add clarification
