### 2024-09-25 3d458c1
paper: add Tc comment
