+++ paper.md
@@ -85,7 +85,7 @@ E, v = system.diagonalize()
 ```
 The results above contain the positive energies $E_n$ and corresponding state vectors $\mathbf{v}_n$ which satisfy the eigenvalue equation $\mathbf{H} \mathbf{v}_n = E_n \mathbf{v}_n$. (Only positive eigenvalues are returned due to the "Nambu doubling" of degrees of freedom in superconducting systems.) Once the eigenvalues and eigenvectors have been obtained, the user can themselves calculate physical properties of interest from these using equations from standard textbooks on the "Bogoliubov–de Gennes" approach to modeling superconductivity [@zhu_bdg_2016]. It is a future goal to incorporate more calculation methods of this kind into the Bodge package itself. Support for matrix diagonalization using GPUs via the CuPy package is also under development.
 
-Examples of physical observables that can be calculated from the eigenvalues and eigenvectors include e.g. the superconducting order parameter, electric currents, and spin currents. These can in turn be used to calculate even more material properties. For instance, the critical current of a Josephson junction is defined as the largest electric current that can flow through it for any phase difference, and the critical temperature of a bulk superconductor is defined as the largest temperature at which the superconducting order parameter remains non-zero. These observables can thus be determined via numerical optimization of the electric current and superconducting order parameter, respectively.
+Examples of physical observables that can be calculated from the eigenvalues and eigenvectors include e.g. the superconducting order parameter, electric currents, and spin currents. These can in turn be used to calculate even more material properties. For instance, the critical current of a Josephson junction is defined as the largest electric current that can flow through it for any phase difference, and the critical temperature of a bulk superconductor is defined as the largest temperature at which the superconducting order parameter remains non-zero. These observables can thus be determined via numerical optimization of the electric current and superconducting order parameter, respectively. For instance, the critical temperature of a superconductor can be efficiently determined using a bisection algorithm [@ouassou_bib_2016; @ouassou_phd_2019].
 
 A modern alternative to matrix diagonalization is a series of algorithms based on Chebyshev expansion of the Hamiltonian matrix [@ouassou_alt_2023; @benfenati_fermi_2022; @covaci_cbdg_2010; @nagai_krylov_2020; @weisse_chebyshev_2006]. These algorithms take advantage of the extreme sparsity of the Hamiltonian matrix, and thus typically provide a significant performance benefit in the limit of very large lattices. One of the main design goals behind Bodge has been to make it trivial to construct sparse representations of the Hamiltonian matrix for this purpose, and it is therefore straight-forward to export the result as e.g. a CSR sparse matrix:
 ```python
