+++ paper.bib
@@ -103,6 +103,17 @@
   pages = {076003},
 }
 
+@article{ouassou_bin_2016,
+  title = {Electric Control of Superconducting Transition through a Spin-Orbit Coupled Interface},
+  author = {Ouassou, Jabir Ali and Bernardo, Angelo Di and Robinson, Jason W. A. and Linder, Jacob},
+  year = {2016},
+  journal = {Scientific Reports},
+  volume = {6},
+  number = {1},
+  pages = {29312},
+  doi = {10.1038/srep29312},
+}
+
 @article{ouassou_dmi_2024,
   title = {{Dzyaloshinskii}--{Moriya} spin--spin interaction from mixed-parity superconductivity},
   url = {http://arxiv.org/abs/2407.07144},
@@ -113,6 +124,15 @@
   month = jul,
   year = {2024},
   note = {arXiv:2407.07144 [cond-mat]},
+}
+
+@book{ouassou_phd_2019,
+  title = {Manipulating Superconductivity in Magnetic Nanostructures in and out of Equilibrium},
+  author = {Ouassou, Jabir Ali},
+  year = {2019},
+  address = {Trondheim, Norway},
+  school = {NTNU},
+  url = {https://pvv.org/~jabirali/academic/phd.pdf}
 }
 
 @article{ouassou_rkky_2024,
