+++ paper.bib
@@ -103,7 +103,7 @@
   pages = {076003},
 }
 
-@article{ouassou_bin_2016,
+@article{ouassou_bisect_2016,
   title = {Electric Control of Superconducting Transition through a Spin-Orbit Coupled Interface},
   author = {Ouassou, Jabir Ali and Bernardo, Angelo Di and Robinson, Jason W. A. and Linder, Jacob},
   year = {2016},
