### 2024-09-25 ba6c1c2
paper: add more sparse info
