### 2024-09-25 07f14fc
paper: fix equation
