### 2024-09-25 53b918c
paper: fix bibliography
