+++ paper.bib
@@ -126,7 +126,7 @@
   note = {arXiv:2407.07144 [cond-mat]},
 }
 
-@book{ouassou_phd_2019,
+@phdthesis{ouassou_phd_2019,
   title = {Manipulating Superconductivity in Magnetic Nanostructures in and out of Equilibrium},
   author = {Ouassou, Jabir Ali},
   year = {2019},
