### 2024-09-25 971b564
paper: improve text
