### 2024-09-25 e1d999d
paper: fix bibliography
