+++ paper.bib
@@ -1,5 +1,5 @@
 @phdthesis{benfenati_fermi_2022,
-  title = {Numerical Solutions to Non-Linear Inhomogeneous Problems in Superconductivity},
+  title = {Numerical solutions to non-linear inhomogeneous problems in superconductivity},
   author = {Benfenati, Andrea Ludovico},
   year = {2022},
   address = {Stockholm, Sweden},
@@ -9,7 +9,7 @@
 }
 
 @article{covaci_cbdg_2010,
-  title = {Efficient Numerical Approach to Inhomogeneous Superconductivity: The Chebyshev-Bogoliubov--de Gennes Method},
+  title = {Efficient numerical approach to inhomogeneous superconductivity: The {Chebyshev}--{Bogoliubov}--de {Gennes} method},
   author = {Covaci, L. and Peeters, F. M. and Berciu, M.},
   year = {2010},
   journal = {Physical Review Letters},
