### 2024-09-25 c599e25
paper: clean bibliography
