+++ paper.bib
@@ -3,7 +3,6 @@
   author = {Benfenati, Andrea Ludovico},
   year = {2022},
   address = {Stockholm, Sweden},
-  langid = {english},
   school = {KTH},
   url = {https://urn.kb.se/resolve?urn=urn:nbn:se:kth:diva-311403},
 }
@@ -14,7 +13,6 @@
   year = {2010},
   journal = {Physical Review Letters},
   volume = {105},
-  number = {16},
   pages = {167006},
   issn = {0031-9007},
   doi = {10.1103/PhysRevLett.105.167006},
@@ -33,7 +31,6 @@
   year = {1994},
   journal = {Physical Review Letters},
   volume = {73},
-  number = {1},
   pages = {122--125},
   issn = {0031-9007},
   doi = {10.1103/PhysRevLett.73.122},
@@ -41,12 +38,9 @@
 
 @article{groth_kwant_2014,
   doi = {10.1088/1367-2630/16/6/063065},
-  url = {https://dx.doi.org/10.1088/1367-2630/16/6/063065},
   year = {2014},
-  month = {jun},
   publisher = {IOP Publishing},
   volume = {16},
-  number = {6},
   pages = {063065},
   author = {Christoph W Groth and Michael Wimmer and Anton R Akhmerov and Xavier Waintal},
   title = {Kwant: a software package for quantum transport},
@@ -58,8 +52,7 @@
   year = {2020},
   title ={pybinding: a Python package for tight-binding calculations},
   journal = {Zenodo},
-  doi = {10.5281/zenodo.4010216},
-  url = {https://doi.org/10.5281/zenodo.4010216}
+  doi = {10.5281/Zenodo.4010216},
 }
 
 @article{nagai_krylov_2020,
@@ -69,36 +62,29 @@
   journaltitle = {Journal of the Physical Society of Japan},
   shortjournal = {J Phys Soc Jpn},
   volume = {89},
-  number = {7},
   pages = {074703},
   issn = {0031-9015},
-  doi = {10.7566/jpsj.89.074703},
+  doi = {10.7566/JPSJ.89.074703},
 }
 
 @article{nagai_ldos_2017,
-  title = {Reduced-Shifted Conjugate-Gradient Method for a Green's Function: Efficient Numerical Approach in a Nano-Structured Superconductor},
+  title = {Reduced-shifted conjugate-gradient method for a {Green's} function: efficient numerical approach in a nano-structured superconductor},
   author = {Nagai, Yuki and Shinohara, Yasushi and Futamura, Yasunori and Sakurai, Tetsuya},
   year = {2017},
   journal = {Journal of the Physical Society of Japan},
   volume = {86},
-  number = {1},
   pages = {014708},
   issn = {0031-9015},
-  doi = {10.7566/jpsj.86.014708},
+  doi = {10.7566/JPSJ.86.014708},
 }
 
 @article{ouassou_alt_2023,
   title = {{DC} {Josephson} effect in altermagnets},
   volume = {131},
   issn = {0031-9007, 1079-7114},
-  url = {https://link.aps.org/doi/10.1103/PhysRevLett.131.076003},
   doi = {10.1103/PhysRevLett.131.076003},
-  language = {en},
-  number = {7},
-  urldate = {2023-08-18},
   journal = {Physical Review Letters},
   author = {Ouassou, Jabir Ali and Brataas, Arne and Linder, Jacob},
-  month = aug,
   year = {2023},
   pages = {076003},
 }
@@ -109,19 +95,15 @@
   year = {2016},
   journal = {Scientific Reports},
   volume = {6},
-  number = {1},
   pages = {29312},
[…]
