### 2024-09-25 1a210f3
paper: add clarification
