### 2024-09-25 jabirali/bodge#4 @jabirali
@yw-fang: I have now implemented your suggestions in the latest version of the manuscript:

- I added a paragraph at the end of the "Statement of Need" section, describing what sparse matrices (e.g. CSR) are and when/why these provide a performance benefit for numerical simulations;
- I added several paragraphs with new discussion and references at the end of the "Examples and Workflows" section, which discuss in more detail different ways Bodge can be used in actual calculations: matrix diagonalization vs. matrix expansion, which observables can be calculated with built-in methods vs. if the user writes some code on their own, and some additional notes on how one can approach e.g. $I_c$ and $T_c$ calculations.

I believe I have now implemented all your requests, and am therefore closing this issue as completed. But of course, feel free to open a new issue if you believe that further improvements are required before proceeding to publication.
