### 2024-09-25 #7134 @jabirali
@yw-fang: I have now revised the manuscript as discussed we discussed in [issue #4](https://github.com/jabirali/bodge/issues/4), and believe I have implemented all the requested changes. Please have a look at the [revised and extended manuscript](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07134/joss.07134/10.21105.joss.07134.pdf).

@phibeck: With these revisions, I think we should be ready to proceed to publication if @yw-fang agrees.
