### 2024-09-25 #7134 @jabirali
@phibeck: Looking at the JOSS instructions for the next steps towards publication:

- I have now [added version tags](https://github.com/jabirali/bodge/tags) to the GitHub repo, including a tagged release [v1.1.2](https://github.com/jabirali/bodge/releases/tag/v1.1.2) for the current version.
- I uploaded Bodge v1.1.2 to Zenodo, and have obtained a DOI identifier for it: [doi:10.5281/zenodo.13839641](https://zenodo.org/doi/10.5281/zenodo.13839641).
