### 2024-09-26 #7134 @phibeck
Hi @yw-fang thank you for confirming and for signing off your review! Many thanks to all reviewers (@mdavezac) for your help with this submission! 🎉

@jabirali the reviewers have recommended the submission for publication! :partying_face: There are a few more steps before we finalize the publication. Thanks for the updates and for sending over the version tag and zenodo DOI. I will still post the regular checklist in my next comment, please make sure to check the remaining items off the list and let me know, thanks.
