### 2024-09-26 #7134 @phibeck
@jabirali at this point could you please:
- make sure that the Zenodo archive title matches that of the JOSS paper?
- make a tagged release with the version number you posted above (it shows as a tag, but isn't a release)
