### 2024-09-26 #7134 @jabirali
Thanks to @phibeck for handling the manuscript, and thanks again to @mdavezac and @yw-fang for helping with the reviews! 🎉

@phibeck: Here are my responses to the points above:

> * Double check authors and affiliations (including ORCIDs)

I have double-checked my name, affiliations, and ORCID in the latest version of the PDF manuscript, and confirm that these are all correct as written.

> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

> * make a tagged release with the version number you posted above (it shows as a tag, but isn't a release)

Thanks for the clarification. I have now made a proper "release" as well:
https://github.com/jabirali/bodge/releases/tag/v1.1.2

>  * Archive the release on Zenodo/figshare/etc and post the DOI here.
>
>  * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

- The Zenodo archive of Bodge v1.1.2 is under this DOI:
[10.5281/zenodo.13839641](https://doi.org/10.5281/zenodo.13839641)
- I have double-checked my name and ORCID in the Zenodo metadata.
- I have now renamed the Zenodo archive title from the software title ("Bodge") to the JOSS paper title ("Bodge: Python package for efficient tight-binding modeling of superconducting nanostructures").

> * Make sure that the license listed for the archive is the same as the software license.

The Zenodo archive and the GitHub project both show that Bodge is published under the "MIT licence" (which is the correct licence).
