### 2024-09-26 03069b4
paper: add doi and reformat isbn
