+++ paper.bib
@@ -20,7 +20,8 @@
 
 @book{degennes_bdg_1966,
   title = {Superconductivity of metals and alloys},
-  isbn = {978-0-7382-0101-6},
+  doi = {10.1201/9780429497032},
+  isbn = {9780429497032},
   author = {{de Gennes}, Pierre Gilles},
   year = {1966},
 }
@@ -151,7 +152,7 @@
 
 @book{zhu_bdg_2016,
   title = {{Bogoliubov}--{de Gennes} method and its applications},
-  isbn = {978-3-319-31312-2},
+  isbn = {9783319313122},
   author = {Zhu, Jian-Xin},
   year = {2016},
   doi = {10.1007/978-3-319-31314-6},
