### 2024-09-26 #7134 @jabirali
> ❌ MISSING DOIs
>
> - 10.1201/9780429497032 may be a valid DOI for title: Superconductivity of metals and alloys

This DOI leads to the correct textbook, so I have pushed a new version of the paper now that simply adds that DOI link to the bibliography.
