### 2024-09-29 fa2f573
feat(wip): re-introduced cuda support + added more eigenvector formats
NOTE: Parts of the code is currently untested and needs serious
testing before being used in "real" calculation. For instance, need to
test that the same eigenvalues are obtained using CUDA, and moreover
that the new "dict" return format is mathematically correct.
