+++ bodge/hamiltonian.py
@@ -170,35 +170,116 @@ class Hamiltonian:
         return Index(k[0, 0])
 
     @typecheck
-    def diagonalize(self) -> tuple[Matrix, Matrix]:
+    def diagonalize(
+        self, cuda=False, format="reshaped"
+    ) -> tuple[Matrix, Matrix] | dict[float, tuple[Matrix, Matrix, Matrix, Matrix]]:
         """Calculate the exact eigenstates of the system via direct diagonalization.
 
-        This calculates and returns the eigenvalues and eigenvectors of the system.
         If you run it as `E, v = system.diagonalize()`, then the eigenvalue `E[n]`
         corresponds to the eigenvector `v[n, :, :]`, where the remaining indices
         of the vector correspond to a position index and a Nambu⊗Spin index. The
         last one means that at a given lattice site `i`, `v[n, i, 0:3]` will give
         you the eigenvector components corresponding to indices {e↑, e↓, h↑, h↓}.
 
-        Please note that this function uses *dense* matrices and thus requires
-        a lot of compute power and memory for large matrices. Many computations
+        This default behavior is referred to as `format="reshaped"` since it is
+        obtained by reshaping the 4N-element long vectors that actually satisfy
+        the eigenvalue equation `H @ v[n] == E[n] * v[n]`. To obtain the "actual"
+        eigenvectors instead, pass the argument `format="raw"` to this method.
+
+        There is one more option available, namely `format="dict"`. This results
+        in the return value being a dictionary. This is structured such that if
+        you run `eig = system.diagonalize(format="dict")`, then you can use
+        `for E, (e_up, e_dn, h_up, h_dn) in eig.items(): ...` to iterate through
+        the eigenvalues and eigenvectors in the system. For each eigenvalue `E`,
+        we then return wave functions `(e_up, e_dn, h_up, h_dn)` that correspond
+        to respectively {e↑, e↓, h↑, h↓} states. These have been reshaped such
+        that `e_up[x, y, z]` corresponds to the spin-up electron wave function
+        at coordinates `(x, y, z)` on the lattice, and so forth. This variant
+        may be a bit slower than the other formats, but is likely the easiest
+        to use as a base for calculating various physical observables in the
+        system due to the close similarity to the analytical BdG equations.
+
+        Please note that this function uses *dense* matrices and thus requires a
+        lot of compute power and memory for large matrices. Many computations
         can be performed using sparse matrix algorithms instead, so if you have
         a lattice with millions of sites I would consider alternative methods.
+
+        Alternatively, if you have an NVIDIA GPU available and install the
+        optional dependency `cupy`, then you can set `cuda=True` when running
+        this function to enable a very fast GPU-accelerated diagonalization.
+        Keep in mind that this only works if the Hamiltonian matrix and all its
+        eigenvectors fit into your available video memory.
         """
         # Convert to a dense matrix.
         H = self.matrix(format="dense")
 
-        # Calculate the relevant eigenvalues and eigenvectors.
-        eigval, eigvec = la.eigh(H, subset_by_value=(0.0, np.inf), overwrite_a=True, driver="evr")
-        eigval = np.array(eigval)
-        eigvec = np.array(eigvec)
+        # Calculate eigenvalues and eigenvectors.
+        if cuda:
+            # GPU-accelerated branch using CuPy.
+            try:
+                # Import libraries.
+                import cupy as cp
+                import cupy.linalg as cla
+
+                # Diagonalize using CUDA.
[…]
