### 2024-09-29 5404895
test(wip): added tests for new diagonaliazation formats
