### 2024-09-29 db47aba
fix: cuda support for diagonalization now working
