### 2024-09-29 b234197
feat: added gpu support to the `free_energy` function
