+++ pyproject.toml
@@ -12,7 +12,7 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
 ]
 dynamic = ["version", "description"]
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 
 dependencies = [
     "beartype",
