### 2024-09-30 3b8cd0a
feat: initial implementation of s-wave gap using diagonalization
Note that there is another implementation on the develop branch using
Fermi matrices instead which can be benchmarked against this one. The
implementation is probably right (for analytical reasons) but needs
way more testing and validation before it's ready for production use.
