### 2024-09-30 #7134 @phibeck
@jabirali here are my only two comments for the manuscript. Please have a look when you have a moment.

- line 47: some words missing in "you just need to subclass Lattice"
- line 187: a missing hyphen in "DzyaloshinskiiMoriya"
