### 2024-09-30 ca9fbf0
paper: implement feedback from @phibeck
