+++ paper.bib
@@ -119,7 +119,7 @@
 }
 
 @article{ouassou_rkky_2024,
-  title = {{RKKY} interaction in triplet superconductors: {Dzyaloshinskii}{Moriya}--type interaction mediated by spin-polarized {Cooper} pairs},
+  title = {{RKKY} interaction in triplet superconductors: {Dzyaloshinskii}---{Moriya}--type interaction mediated by spin-polarized {Cooper} pairs},
   volume = {109},
   issn = {2469-9950, 2469-9969},
   shorttitle = {{RKKY} interaction in triplet superconductors},
