### 2024-09-30 #7134 @jabirali
@phibeck: Thanks a lot for going over the manuscript! I have now implemented your two corrections in the newest version of the paper:

- Clarification of "you just need to subclass Lattice":

- Correction of the bibliography item:
