### 2024-10-08 #7134 @kyleniemeyer
Hi @jabirali, I made some typographic edits in https://github.com/jabirali/bodge/pull/5. Can you review and merge this?
