+++ paper.md
@@ -27,13 +27,13 @@ bibliography: paper.bib
 
 [Bodge](https://github.com/jabirali/bodge) is a Python package for constructing *large-scale real-space tight-binding models* for calculations in condensed matter physics. "Large-scale" means that it should remain performant even for lattices with millions of atoms, and "real-space" means that the model is formulated in terms of individual lattice sites and not in momentum space, for example.
 
-Although general tight-binding models can be constructed with this package, the main focus is on the Bogoliubov-DeGennes ("BoDGe") Hamiltonian used to model superconductivity in the clean limit [@zhu_bdg_2016; @degennes_bdg_1966]. The package is designed to be easy to use, flexible, and extensible-–-and very few lines of code are required to model heterostructures containing, e.g., conventional and unconventional superconductors, ferromagnets and antiferromagnets, altermagnetism, and spin-orbit coupling.
+Although general tight-binding models can be constructed with this package, the main focus is on the Bogoliubov--De Gennes ("BoDGe") Hamiltonian used to model superconductivity in the clean limit [@zhu_bdg_2016; @degennes_bdg_1966]. The package is designed to be easy to use, flexible, and extensible---and very few lines of code are required to model heterostructures containing, e.g., conventional and unconventional superconductors, ferromagnets and antiferromagnets, altermagnetism, and spin-orbit coupling.
 
 In other words: If you want a lattice model for superconducting nanostructures, and want something that is computationally efficient yet easy to use, Bodge should be a good choice.
 
 # Statement of need
 
-In condensed matter physics, a standard methodology for modeling materials is the *tight-binding model*. In the context of electronic systems (e.g., metals), the electrons in such a model typically "live" at one atomic site, but from time to time "hop" over to neighboring atoms. By including a spin structure as well in this formalism-–-meaning that we keep track of what spins each electron has, and whether the spins "flip" during various interactions that are permitted on this lattice-–-we can model a wide variety of physical phenomena including superconductivity and magnetism. Mathematically, this is often expressed in the language of quantum field theory: We define one operator $c^\dagger_{i\sigma}$ that "puts" an electron with spin $\sigma \in \{\uparrow, \downarrow\}$ on an atomic site with some index $i$, and another operator $c_{i\sigma}$ that "removes" a corresponding electron. The Hamiltonian operator $\mathcal{H}$ of the system is then constructed out of these electron operators-–-and this can in turn be used to calculate, e.g., the ground-state energy, electric currents, superconducting order parameters, and other relevant material properties.
+In condensed matter physics, a standard methodology for modeling materials is the *tight-binding model*. In the context of electronic systems (e.g., metals), the electrons in such a model typically "live" at one atomic site, but from time to time "hop" over to neighboring atoms. By including a spin structure as well in this formalism---meaning that we keep track of what spins each electron has, and whether the spins "flip" during various interactions that are permitted on this lattice---we can model a wide variety of physical phenomena including superconductivity and magnetism. Mathematically, this is often expressed in the language of quantum field theory: We define one operator $c^\dagger_{i\sigma}$ that "puts" an electron with spin $\sigma \in \{\uparrow, \downarrow\}$ on an atomic site with some index $i$, and another operator $c_{i\sigma}$ that "removes" a corresponding electron. The Hamiltonian operator $\mathcal{H}$ of the system is then constructed out of these electron operators---and this can in turn be used to calculate, e.g., the ground-state energy, electric currents, superconducting order parameters, and other relevant material properties.
 
 To do anything useful with that Hamiltonian *on a computer*, however, you typically have to translate it to a matrix form. This is where Bodge enters the picture:
 
[…]
