### 2024-10-09 #7134 @jabirali
@kyleniemeyer: Thanks for the revision! I have merged this now.

Most of the revisions look good. There was however one minor issue in the PR: I had originally formatted the MarkDown file using Unicode symbols for en-dashes (–) and em-dashes (—), and your PR adds hyphens around these dashes instead of replacing them. Mixes like -–- (hyphen/en-dash/hyphen) are not understood by the compiler which causes some artifacts in the PDF version.

I have now pushed a fix for the dash formatting in the MarkDown.
